Add sorting of fractions based on index and organize folder structure
