+++ client/include/fraction.h
@@ -26,4 +26,5 @@ int fraction_parse(char *data, size_t size, fraction_t *fraction);
 int check_magic(uint32_t data);
 void print_fraction(fraction_t fraction);
 void fraction_free(fraction_t *fraction);
+int compare_fractions(const void* a, const void* b);
 #endif
