+++ client/src/http.c
@@ -1,4 +1,4 @@
-#include "http.h"
+#include "../include/http.h"
 
 #define HTTP_BUFFER_SIZE 1024
